Adopt new `pooch`-based versioning system approach (#43)
### Description

The modifications here completely change the way that testing data is managed in order to provide a versioning system as well as reduce the number of remote requests to GitHub.

Projects that rely on `raven-testdata` will now use `pooch` alongside file registries to track the file and checksums, as well as manage the caches. This repo will also now be using tagged commits to ensure that testing data is kept up-to-date while at the same time facilitating easier backtracking if needed.

### Changes

- All `md5` files have been removed. 
- The `README.md` has been updated to detail the new process for fetching data with `pooch`
- The `report_check_sums.py` script now creates new registries as well as updates the registry table in the `README.md`
- The return characters in the text files have been converted to `LF` to ensure that checksums remain the same when opened on Linux and macOS; lots of redundant whitespace has also been removed from tables.

### What's next

Once this PR is approved, `main` will be tagged using calendar versioning as the first stable version. Once all projects reliant on `raven-testdata` have been updated to the new system, the default branch will change from `master` to `main`.

The `master` branch will be kept for backwards-compatibility.

See also: https://github.com/CSHS-CWRA/RavenPy/pull/513